Implement initial version of bellman ford arbitrage path finding. The resulting negative cycle with always be the largest, but might not include the specified source currency. This change requires updating the tests.
TODO: As a result of the way bellman-ford works, nested cycles are not
allowed. The result is that a currency cannot be traded in more than two
places at once. It will enable more opportunities if we modify the graph
structure to treat a vertex as a pair of currency,exchange instead of
just currency. Then it will be possible to trade a currency on multiple
exchanges at the same time. This refactor will be a pain though, defer
until later.